pyproject.toml
[tool.poetry]
name = "bcda-claims-sandbox"
version = "0.1.0"
description = ""
authors = ["lwarrenburg <lindsay.a.warrenburg@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.scripts]
bcda = 'src.main:cli'

[tool.poetry.dependencies]
python = "^3.10,<3.13"
dbt-snowflake = "^1.6.2"
click = "^8.1.7"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"
snowflake-sqlalchemy = "^1.5.0"
pandas = ">=1.0.0,<2.1.0"
snowflake-connector-python = {extras = ["pandas"], version = "^3.3.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
